Add the manual script
This is not very interesting, as it's not how the IBus daemon
loads our engine. However, it's good to have it for tests.

And it seems some stuff loads engines without the IBus daemon.

It's not correct at the moment though, as it only loads the
component, but doesn't add the engine. We'll figure something
out to load them from the XML, to avoid duplicating the code.