Use NoDisplay=true instead of NotShowIn
The goal is to hide the setup launchers from the menus.

Both work fine for that, but NotShowIn=GNOME; has an undesirable
effet: when running the setup application, its label in the GNOME
Shell application menu is set to the name of the executable.

NoDisplay=true achieves the same goal, still showing the nice name
for the user.