Split out the two engines
IBus loads the component the first time (for each session)
that one of the engines in that component is used.

I'd rather not have both Cangjie and Quick constantly loaded
together, even though the vast majority of people will only
use one or the other, but not both.

This commit switches the engines in two distinct XML files.