Factorize the handling of the IBus config
This was all defined in the setup UI. But at some point, the
engine itself will need to read its settings.

So it makes sense to have that in a dedicated Config class.

It is read-only by default and must be explicitly turned
read-write at instanciation time. This will prevent the engine
from changing its own config, which really shouldn't ever happen.
Only settings UI should change the settings.

The setup UI has been ported to it.