Implement ibus_cangjie.setup.Setup.on_value_changed
So I got that function from whichever engine I used as a template for
the settings UI (ibus-tmpl? ibus-hangul? can't remember), and I didn't
really know what it was for.

That's is why I had left some print statements in there so I'd see
something happening when it was used.

I finally figured it out, so here it is.