Separate the `update_lookup_table` method
The `update` method did a lot of stuff, and I'm about to add some more
to it.

I'd prefer have smaller, simpler methods, each taking care of its own
thing. Plus, the parent class already has a bunch of simple methods for
each thing to update (preedit text, auxiliary text, lookup table), so we
might as well just build upon that.

This is the first step towards that aim, and the code related to the
lookup table is now handled on its own.