config: Ensure default values
Currently, we were ensuring the default config values when
running the setup UI for the first time.

This is wrong, because then the engine won't know the default
values, or it will have to ensure them itself on its own first
run if the setup UI is never launched.

With this commit, the config.Config class itself will ensure
the default value on instanciation, which happens both when
running the setup UI and the engine.

The setup UI code had to be adapted to the new way as well.