Great rename
So far, I had been using the terms "preedit" or "pre-edit" to talk about
the currently inputted English characters which haven't yet been
committed as a Chinese character.

This became confusing when I started displaying this text as the
"auxiliary text" instead of the "preedit text".

It turns out that I was confused. The "preedit" and "auxiliary" texts
are about where they are shown, not what they contain.

Because neither Cangjie nor Quick use "preedit" (everything is displayed
as "auxiliary"), I shouldn't have called the current input like that.