Modify the update_current_input API
We will need to display the radicals for each inputted English
character.

To avoid iterating over the new current input every time it is modified
and running CangJie.getCharacters() over each, we need a way to only
append new characters to the current input, and run
CangJie.getCharacters() over just that.

Hence the new function signature.