cangjie: Input after space on multiple candidates
When pressing space, the user sometimes gets multiple candidates.

When that happens, if the user then goes on pressing an input key, we
shouldn't consider that as the continuation of the previous input.

Instead, we should commit the first candidate, and start a new input
with that new key.

fixes #8