Limit ourselves to Chinese
We declare ourselves as handling Chinese in the XML component file, and
(as far as I know) incredibly few people use Cangjie/Quick to input
Japanese.

So let's limit what we do a bit for now, and try to make it rock.

If people really want Japanese, we can add it later on.