setup: Pretend to be a GNOME Control Center dialog
In GNOME, our preferences dialog is run through GNOME Control Center,
not directly (because we have NotShowIn=GNOME in our desktop file).

As such, it is important to pretend that we are a native dialog from
GNOME Control Center, so that the user doesn't feel like she is being
presented with various pieces: she should see a consistent whole.

Of course, this only applies when we are run by GNOME Control Center. It
does not change anything for users launching our preferences on other
desktop environments (e.g Unity or KDE) or running the command-line
executable directly.

This fixes bug #15.