tests: Monkey patch the environment
We don't want to connect tot he actual bus to run our unit tests.

This allows us to completely instrument our engine class and test its
behaviour.

The test-specific attributes in our monkey-patched classes should all
start with "_mock_", for example:
    MockEngine._mock_auxiliary_text

This should help make sure we don't create conflicts with the actual
attributes of the objects we are monkey-patching.