cangjie: Properly handle star as first character
In Cangjie, the asterisk (*) can only be used as a wildcard if it is
**surrounded** by text, that is, it needs **at least** one character
before and one character after.

As such, if the first thing the user types is an asterisk, we shouldn't
consider it as a wildcard, but as a plain character to type directly.

This commit introduces two new unit tests which make sure we will not
regress on this feature.