Make the property label translatable
This solution sucks, because we're hardcoding the gettext package in
src/config.py.

The only alternative I could think of was to generate src/config.py at
build time, as it used to be.

But then that was making unit testing impossibly hard.

So if I have to choose between (i18n + unit tests) and "code purity",
the former appears **much** more appealing.

At least I'm adding a FIXME item and a note in configure.ac.