Stop bundling pycanberra
It came to my attention that pycanberra is LGPLv2 only, while IBus
Cangjie is LGPLv3+.

The mix of the two, bundled together might be problematic, so to avoid
any issue, dropping our copy of pycanberra is the easiest way out.

The code was already handling errors to play sound anyway (no sound
would be played, so let's generalize that to the case where pycanberra
is just not available.