Relicense to GPLv3+
IBus Cangjie is not really a library, so the going with LGPL instead
of plain GPL doesn't make much sense in retrospect.

I asked Benau (the only contributor other than myself) and he agreed
in an email.

So all copyright holders agreed, and from now on, IBus Cangjie is
distributed under the GPLv3+.