README: Move to markdown, and misc improvements
Just like we did for libcangjie2 and pycangjie, we'll reuse this README
for the ibus-cangjie home page at http://cangjians.github.io, so it
makes sense to have it in Markdown rather than rst.

This also drops mention of the AUTHORS and INSTALL.rst files, as they
were confusing people.