tests: Don't print the test logs
It seems that the log file started being generated with more recent
versions of Automake.

The Travis CI server is running Ubuntu 12.04, where that file does not
get generated apparently.

In any case, this was just to make my life slightly easier when I was
working on the tests, but it's not that useful as the file can be
printed manually anyway.