Port the engine configuration to GSettings
IBus upstream will deprecate IBusConfig, and recommends engines to
migrate to GSettings instead.

Given that we are breaking the configuration options laready between 1.0
and 2.0, it seems like the good moment to make that migration, so we
don't break user configurations again in the future.

It does require quite a lot of change, but overall it seems like a net
win, as it allows us to remove quite a bit of code.

Fixes #16