tests: Assume pycanberra is always there
Thanks to the previous commit, we can now assume that pycanberra is
always available for the unit tests.

That makes it much easier to mock the pycanberra stuff, and it removes
the need to fiddle with an environment variable to let the tests know
whether pycanberra is here or not.