i18n: Pull in the latest pt_BR translation
I don't think that "Quick" should be translated into Portuguese when it
is used as the name of the input method. It is a name, not an adjective.

We do translate it into Chinese, but that is because Chinese actually
has a real name for that input method.

The Portuguese language however doesn't know about Chinese input method,
and as such we shouldn't translate "Quick", just like we don't translate
"Cangjie".