Work around the candidate popup position issue
Some applications (most notably Firefox) fail to correctly position the
candidate popup.

This is because we don't use a pre-edit text at all, and it seems those
applications (or maybe the toolkits?) get confused by the absence of a
pre-edit text. :(

This is a gross hack. The real issue needs to be fixed in the
apps/toolkits, but we need to do something in the meantime, as this is
just too much of a bad UX issue for us.

The idea comes from FrÃdÃric Péter.

Fixes #60